refactor(components): migrate final 3 SFCs to h() render functions
배치 3으로 마지막 SFC 3개를 전환해 src 에 .vue 파일이 남지 않는다.

- KrdsTooltip (216줄): v-show 를 withDirectives + vShow 로 전환
- KrdsPagination (266줄): <component :is> 를 동적 태그 h() 로,
  커스텀 디렉티브 v-sr-only 를 withDirectives 로 전환
- KrdsDateInput (550줄): 캘린더 그리드/드롭다운을 렌더 헬퍼로 분리,
  KrdsButton 자식은 슬롯 객체로 전달

defineComponent<Props> 는 withDefaults 와 달리 optional 프롭을 런타임
기본값으로 좁혀주지 않는다. Pagination 의 min/pageRange 는 기본값을
상수로 뽑아 props 선언과 setup 이 같은 값을 참조하도록 했다.
