+++ src/components/KrdsPagination/index.ts
@@ -2,8 +2,8 @@
  * KRDS Pagination 컴포넌트
  */
 
-import KrdsPagination from './KrdsPagination.vue'
-import type { KrdsPaginationProps, KrdsPaginationEmits } from './KrdsPagination.vue'
+import KrdsPagination from './KrdsPagination'
+import type { KrdsPaginationProps, KrdsPaginationEmits } from './KrdsPagination'
 
 export { KrdsPagination }
 export type { KrdsPaginationProps, KrdsPaginationEmits }
